:recycle: refactor: separate directories according to kind
