+++ src/ui/page/LoginPage.ts
@@ -1,9 +1,9 @@
-import { CustomElement, Notification } from './CustomElement';
-import TEMPLATE from '../templates';
-import { addEvent, emit, $, showSnackbar } from '../utils';
-import Authentication from '../domain/Authentication';
-import { historyRouterPush } from '../router';
-import { BASE_URL, CUSTOM_EVENT, ELEMENT_KEY } from '../constants';
+import { CustomElement, Notification } from '../CustomElement';
+import TEMPLATE from '../../templates';
+import { addEvent, emit, $, showSnackbar } from '../../utils';
+import Authentication from '../../domain/Authentication';
+import { historyRouterPush } from '../../router';
+import { BASE_URL, CUSTOM_EVENT, ELEMENT_KEY } from '../../constants';
 
 class LoginPage extends CustomElement {
   connectedCallback() {
