+++ src/ui/page/SignupPage.ts
@@ -1,8 +1,8 @@
-import { CustomElement, Notification } from './CustomElement';
-import TEMPLATE from '../templates';
-import { addEvent, emit, showSnackbar, $ } from '../utils';
-import Authentication from '../domain/Authentication';
-import { CUSTOM_EVENT, ELEMENT_KEY } from '../constants';
+import { CustomElement, Notification } from '../CustomElement';
+import TEMPLATE from '../../templates';
+import { addEvent, emit, showSnackbar, $ } from '../../utils';
+import Authentication from '../../domain/Authentication';
+import { CUSTOM_EVENT, ELEMENT_KEY } from '../../constants';
 
 class SignupPage extends CustomElement {
   connectedCallback() {
